Short-circuit if the license file already exists.
The license file might already exist because it was present in the repo, because it's already present from an sdist, or it's there from a prior invocation. In any case, only download the resolved license when a license file is not already present.

Ref jaraco/skeleton#178.

+++ setuptools.py
@@ -9,8 +9,10 @@ def _finalize_license_files(dist):
     Resolve the license expression into a license file.
     """
     license = pathlib.Path('LICENSE')
-    license.write_text(resolve(dist.metadata.license_expression))
     dist.metadata.license_files = [str(license)]
+    if license.exists():
+        return
+    license.write_text(resolve(dist.metadata.license_expression))
 
 
 def inject(dist):
